feat(reportOfInactiveWorkplaces) renamed the find<workplaceTpye> files to set<workplaceType> due to no longer destructuring the files
