+++ server/reports/inactive-workplaces/index.js
@@ -1,6 +1,6 @@
 const excelUtils = require('../../utils/excelUtils');
 
-const findParentWorkplaces = require('../../services/email-campaigns/inactive-workplaces/findParentWorkplaces');
+const setParentWorkplaces = require('../../services/email-campaigns/inactive-workplaces/setParentWorkplaces');
 
 const { generateInactiveWorkplacesTab } = require('./workplaces');
 const { generateParentWorkplaceTab } = require('./parents');
@@ -8,7 +8,7 @@ const { generateSubsidaryWorkplaceTab } = require('./subsidiaries');
 const { generateInactiveWorkplacesForDeletionTab } = require('./deleteInactiveWorkplace');
 
 const generateInactiveWorkplacesReport = async (workbook) => {
-  const parentWorkplaces = await findParentWorkplaces.findParentWorkplaces();
+  const parentWorkplaces = await setParentWorkplaces.findParentWorkplaces();
 
   await generateInactiveWorkplacesTab(workbook);
   await generateParentWorkplaceTab(workbook, parentWorkplaces);
