+++ server/services/email-campaigns/inactive-workplaces/setInactiveWorkplaces.js
