+++ server/services/email-campaigns/inactive-workplaces/setInactiveWorkplacesForDeletion.js
